+++ app/db/auth.js
@@ -1,0 +1,4 @@
+import { getAuth } from 'firebase/auth';
+import { app } from './firebase-setup';
+
+const auth = getAuth(app);
